Filter the commits better when resolving tags and branches.
Was sometimes including commits that added a file on the trunk, which were then backported to the branch - the initial revision on the trunk was counted as an ancestor of the branch revision - if other files were included in that commit they would also be applied.

This approach works out the "branch path" of the branch that the tag is on and only considers commits along this path, so excludes any commit on the trunk made after the branch is made.